@@ -1,31 +1,34 @@
 {
   "name": "unused-webpack-plugin",
-  "version": "2.4.0",
+  "version": "2.4.1",
   "description": "A webpack plugin to find unused modules/source files",
   "main": "index.js",
-  "repository": "https://github.com/MatthieuLemoine/unused-webpack-plugin",
-  "author": "MatthieuLemoine",
+  "repository": "https://github.com/Better-IRIS/unused-webpack-plugin",
+  "author": "IRIS",
   "license": "MIT",
   "scripts": {
-    "test": "jest",
+    "test": "yarn node --experimental-vm-modules $(yarn bin jest)",
+    "test:coverage": "yarn node --experimental-vm-modules $(yarn bin jest --coverage)",
     "lint": "eslint .",
-    "prettify": "prettier-eslint --write \"**/*.js*\" --list-different",
-    "pretty-check": "prettier-eslint \"**/*.js*\" --list-different",
+    "prettify": "prettier-eslint --write \"**/*.ts*\" --list-different",
+    "pretty-check": "prettier-eslint \"**/*.ts*\" --list-different",
     "check-version": "node scripts/check-version.js"
   },
   "devDependencies": {
-    "eslint": "^5.6.1",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.14.0",
-    "husky": "^1.1.1",
-    "jest": "^23.6.0",
-    "lint-staged": "^7.3.0",
-    "prettier": "^1.14.3",
-    "prettier-eslint-cli": "^4.7.1"
+    "@types/jest": "^29.4.0",
+    "eslint": "^8.33.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.27.5",
+    "husky": "^8.0.3",
+    "jest": "^29.4.1",
+    "lint-staged": "^13.1.0",
+    "prettier": "^2.8.3",
+    "prettier-eslint-cli": "^7.1.0"
   },
   "dependencies": {
-    "chalk": "^2.1.0",
-    "deglob": "^3.1.0"
+    "chalk": "^5.2.0",
+    "deglob": "^4.0.1",
+    "typescript": "^4.9.5"
   },
   "husky": {
     "hooks": {
@@ -49,6 +52,7 @@
     "testPathIgnorePatterns": [
       "/node_modules/",
       "<rootDir>/examples/"
-    ]
+    ],
+    "transform": {}
   }
 }
