@@ -1,27 +1,53 @@
+const { readFileSync, writeFile } =require('fs');
 const path = require('path');
 const chalk = require('chalk');
 const { searchFiles } = require('./lib/utils');
 
-function UnusedPlugin(options) {
+function UnusedPlugin(
+  this: {
+    sourceDirectories: string[];
+    exclude: string[];
+    root: 'string';
+    failOnUnused: boolean;
+    useGitIgnore: boolean;
+    outputFilePath: string;
+  },
+  options: {
+    directories: string[];
+    exclude: string[];
+    root: 'string';
+    failOnUnused: boolean;
+    useGitIgnore: boolean;
+    outputFilePath: string;
+  },
+) {
   this.sourceDirectories = options.directories || [];
   this.exclude = options.exclude || [];
   this.root = options.root;
   this.failOnUnused = options.failOnUnused || false;
   this.useGitIgnore = options.useGitIgnore || true;
+  this.outputFilePath = options.outputFilePath;
 }
 
-UnusedPlugin.prototype.apply = function apply(compiler) {
+UnusedPlugin.prototype.apply = function apply(this: {
+  sourceDirectories: string[];
+  exclude: string[];
+  root: 'string';
+  failOnUnused: boolean;
+  useGitIgnore: boolean;
+  outputFilePath: string;
+}, compiler) {
   const checkUnused = (compilation, callback) => {
     // Files used by Webpack during compilation
     const usedModules = Array.from(compilation.fileDependencies)
-      .filter(file => this.sourceDirectories.some(dir => file.indexOf(dir) !== -1))
+      .filter((file) => this.sourceDirectories.some((dir) => file.indexOf(dir) !== -1))
       .reduce((obj, item) => Object.assign(obj, { [item]: true }), {});
     // Go through sourceDirectories to find all source files
     Promise.all(
-      this.sourceDirectories.map(directory => searchFiles(directory, this.exclude, this.useGitIgnore)),
+      this.sourceDirectories.map((directory) => searchFiles(directory, this.exclude, this.useGitIgnore)),
     )
       // Find unused source files
-      .then(files => files.map(array => array.filter(file => !usedModules[file])))
+      .then((files) => files.map((array) => array.filter((file) => !usedModules[file])))
       .then(display.bind(this))
       .then(continueOrFail.bind(this, this.failOnUnused, compilation))
       .then(callback);
@@ -35,7 +61,7 @@ UnusedPlugin.prototype.apply = function apply(compiler) {
   }
 };
 
-module.exports = UnusedPlugin;
+export default UnusedPlugin;
 
 function continueOrFail(failOnUnused, compilation, allFiles) {
   if (allFiles && allFiles.length > 0) {
@@ -55,6 +81,11 @@ function display(filesByDirectory) {
   if (!allFiles.length) {
     return [];
   }
+
+  let outputString = `## Unused Files
+${allFiles.length} unused source files found.
+`;
+
   process.stdout.write('\n');
   process.stdout.write(chalk.green('\n*** Unused Plugin ***\n'));
   process.stdout.write(
@@ -63,15 +94,30 @@ function display(filesByDirectory) {
   filesByDirectory.forEach((files, index) => {
     if (files.length === 0) return;
     const directory = this.sourceDirectories[index];
-    const relative = this.root
-      ? path.relative(this.root, directory)
-      : directory;
+    const relative = this.root ? path.relative(this.root, directory) : directory;
+
     process.stdout.write(chalk.blue(`\n● ${relative}\n`));
-    files.forEach(file => process.stdout.write(
-      chalk.yellow(`    • ${path.relative(directory, file)}\n`),
-    ));
+    outputString += `\n### ${relative}\n`;
+
+    files.forEach((file) => {
+      process.stdout.write(
+        chalk.yellow(`    • ${path.relative(directory, file)}\n`),
+      );
+      outputString += `  - ${path.relative(directory, file)}\n`;
+    });
   });
   process.stdout.write(chalk.green('\n*** Unused Plugin ***\n\n'));
+
+  if (this.outputFilePath) {
+    writeFile(this.outputFilePath, outputString, 'utf8', (err) => {
+      if (err) console.log(err);
+      else {
+        console.log('File written successfully\n');
+        console.log('The written has the following contents:');
+        console.log(readFileSync('books.txt', 'utf8'));
+      }
+    });
+  }
 
   return allFiles;
 }
