@@ -1,5 +1,5 @@
 const path = require('path');
-const UnusedWebpackPlugin = require('../');
+const UnusedWebpackPlugin = require('..');
 
 module.exports = {
   target: 'node',
@@ -20,6 +20,7 @@ module.exports = {
       // Root directory (optional)
       root: __dirname,
       failOnUnused: false,
+      outputFilePath: path.resolve(__dirname, 'UNUSED_FILES.MD'),
     }),
   ],
 };
